+++ src/pvoutput/pvoutput.py
@@ -1,16 +1,15 @@
-""" AsyncIO interface to the PVOutput API """
+""" provides an class for uploading to the PVOutput API """
 
 import datetime
-import re
 
-import asyncio
-import aiohttp
+import requests
 
 from .parameters import *
+
 from . import utils
 
 
-class AsyncPVOutput:
+class PVOutput:
     """ This class provides an interface to the pvoutput.org API """
 
     validate_data = utils.validate_data
@@ -22,9 +21,10 @@ class AsyncPVOutput:
         systemid: int,
         donation_made: bool = False,
         stats_period: int = 5,
-        session=None,
     ):
-        """Startup phase
+        """Setup code
+
+
         ```
         :param apikey: API key (read or write)
         :type apikey: str
@@ -34,10 +34,6 @@ class AsyncPVOutput:
         :type donation_made: bool
         ```
         """
-        self.session = session
-        if not session:
-            self.session = aiohttp.ClientSession()
-
         if not isinstance(systemid, int):
             raise TypeError("systemid should be int")
         if not isinstance(apikey, str):
@@ -60,9 +56,7 @@ class AsyncPVOutput:
         }
         return headers
 
-    async def _call(
-        self, endpoint, data=None, params=None, headers=False, method: str = "POST"
-    ):
+    def _call(self, endpoint, data=None, params=None, headers=False, method="POST"):
         """Makes a call to a URL endpoint with the data/headers/method you require.
 
         :param endpoint: The URL to call
@@ -84,14 +78,10 @@ class AsyncPVOutput:
         :raises ValueError: if the call throws a HTTP 400 error.
         :raises requests.exception: if method throws an exception.
         """
-        if not method:
-            method = self.session.post
-
         # always need the base headers
         if not headers:
             headers = self._headers()
         # TODO: type checking on call
-
         if method == "POST" and data and isinstance(data, dict) is False:
             raise TypeError(f"data should be a dict, got {str(type(data))}")
         if method == "GET" and params and isinstance(params, dict) is False:
@@ -100,24 +90,20 @@ class AsyncPVOutput:
             raise TypeError(f"headers should be a dict, got {str(type(headers))}")
         # TODO: learn if I can dynamically send thing, is that **args?
         if method == "GET":
-            response = await self.session.get(
-                endpoint, data=data, headers=headers, params=params
-            )
+            response = requests.get(endpoint, data=data, headers=headers, params=params)
         else:
-            response = await self.session.post(endpoint, data=data, headers=headers)
+            response = requests.post(endpoint, data=data, headers=headers)
 
-        if response.status == 400:
+        if response.status_code == 400:
             # TODO: work out how to get the specific response and provide useful answers
-            raise ValueError(f"HTTP400: {(await response.text()).strip()}")
+            raise ValueError(f"HTTP400: {response.text.strip()}")
         # likely errors - https://pvoutput.org/help/api_specification.html#error-messages
         response.raise_for_status()
         return response
 
-    async def check_rate_limit(self):
+    def check_rate_limit(self):
         """
-        Makes a call to the site, checking if you have hit the rate limit. Check the `documentation`_ for details.
-
-        .. _documentation: https://pvoutput.org/help.html#api-ratelimit
+        Makes a call to the site, checking if you have hit the rate limit. Check the [documentation](https://pvoutput.org/help/api_specification.html#rate-limits)
 
         :returns: the headers relating to the rate limit.
         :rtype: dict
@@ -128,12 +114,11 @@ class AsyncPVOutput:
 
         url, method = utils.URLS["getsystem"]
 
-        response = await self._call(url, {}, headers=headers, method=method)
+        response = self._call(url, {}, headers=headers, method=method)
         retval = utils.get_rate_limit_header(response)
-
         return retval
 
-    async def addstatus(self, data: dict):
+    def addstatus(self, data: dict):
         """
         The Add Status service accepts live output data at the status interval (5 to 15 minutes) configured for the system.
 
@@ -159,7 +144,7 @@ class AsyncPVOutput:
 
         url, method = utils.URLS["addstatus"]
 
-        return await self._call(endpoint=url, data=data, method=method)
+        return self._call(endpoint=url, data=data, method=method)
 
     # def addoutput(self, data: dict):
     #     """ The Add Output service uploads end of day output information. It allows all of the information provided on the Add Output page to be uploaded.
@@ -175,7 +160,7 @@ class AsyncPVOutput:
     #     # self.validate_data(data, ADDOUTPUT_PARAMETERS)
     #     # self._call(endpoint="https://pvoutput.org/service/r2/addoutput.jsp", data=data)
 
-    async def delete_status(self, date_val: datetime.datetime.date, time_val=None):
+    def delete_status(self, date_val: datetime.datetime.date, time_val=None):
         """deletes a given status, based on the provided parameters
         needs a datetime() object
         set the hours/minutes to non-zero to delete a specific time
@@ -217,7 +202,8 @@ class AsyncPVOutput:
         )
         return response
 
-    async def getstatus(self) -> dict:
+    # pylint: disable=too-many-locals
+    def getstatus(self) -> dict:
         """
         Makes a call to the API and gets the last update.
 
@@ -230,12 +216,11 @@ class AsyncPVOutput:
         data = {}
         if self.donation_made:
             url = f"{url}?ext=1&sid={self.systemid}"
-        response = await self._call(endpoint=url, data=data, method="GET")
-
+        response = self._call(endpoint=url, data=data, method="GET")
+        response.raise_for_status()
         # grab all the things
-        text = await response.text()
-
-        responsedata, extras = utils.responsedata_to_response(text.split(","))
+        # pylint: disable=invalid-name
+        responsedata, extras = utils.responsedata_to_response(response.text.split(","))
 
         # if we're fancy, we get more data
         if extras:
@@ -245,15 +230,14 @@ class AsyncPVOutput:
                 )
         return responsedata
 
-    async def register_notification(self, appid: str, url: str, alerttype: int):
+    def register_notification(self, appid: str, url: str, alerttype: int):
         """
         The Register Notification Service allows a third party application to receive PVOutput alert callbacks via a HTTP end point.
 
-        [API Documentation](https://pvoutput.org/help.html#api-registernotification)
+        `API Documentation`_
 
-        All parameters are mandatory:
+        All parameters are mandatory
 
-        ```
         :param appid: Application ID (eg: example.app.id)
         :type appid: str (maxlen: 100)
 
@@ -263,6 +247,7 @@ class AsyncPVOutput:
         :param type: Alert Type (See list below)
         :type type: int
 
+        .. _API Documentation: https://pvoutput.org/help.html#api-registernotification
 
         Type list:
 
@@ -287,7 +272,6 @@ class AsyncPVOutput:
         23      High Net Power Alert
         24      Low Net Power Alert
         =====   ====
-        ```
         """
         # TODO: Find out if HTTPS is supported for Callback URLs
         # TODO: validation of types, is this the best way?
@@ -302,7 +286,7 @@ class AsyncPVOutput:
             )
         if len(appid) > 100:
             raise ValueError(
-                f"Length of appid can't be longer than 100 chars - was {len(appid)}"
+                f"Length of appid can't be longer than 150 chars - was {len(appid)}"
             )
 
         if not isinstance(alerttype, int):
@@ -312,5 +296,5 @@ class AsyncPVOutput:
         # TODO: urlencode the callback URL
 
         call_url = f"https://pvoutput.org/service/r2/registernotification.jsp?appid={appid}&type={alerttype}&url={url}"
-        response = await (await self._call(endpoint=call_url, method="GET")).text()
+        response = self._call(endpoint=call_url, method="GET")
         return response
